Updated files due to linter
