+++ augbuilder/augmentation.py
@@ -1,5 +1,4 @@
 import albumentations
-import autopep8
 import streamlit as st
 
 from elements import (
@@ -181,62 +180,4 @@ def dict_update(
             current_choice,
             augmentations,
             session_state,
-            
         )
-
-
-def build_string():
-    result_text = ''
-    for augm in list(aug_dict.keys()):             
-        result_text += '{0}:\n'.format(augm)
-        key_result = ''
-        if aug_dict[augm]:
-            for elem in aug_dict[augm]:  # noqa: WPS528
-                str_temp = '\t{0}: {1}\n'.format(
-                    elem,
-                    aug_dict[augm][elem],
-                )
-                key_result += str_temp
-        result_text += key_result
-    return result_text
-
-
-def build_code_substring(iterable):
-    last_iter = 0
-    res = ''
-    for imp in iterable:
-        if last_iter < len(list(iterable)) - 1:
-            res += str(imp) + ',\n'
-        else:
-            res += str(imp) + ','
-        last_iter += 1
-    return res
-
-
-def build_code():
-    imports = list(aug_dict.keys())
-
-    res_imports = build_code_substring(imports)
-    composes = build_code_substring(apply_changes(aug_dict))
-    pytorch2tensor = ''
-    if st.sidebar.checkbox('add ToTensorv2()'):  
-        pytorch2tensor = 'from albumentations.pytorch import ToTensorV2'
-        composes += '\nToTensorV2(),'
-
-    result = """{pytorch2tensor}
-    from albumentations import (
-        Compose,
-        {imports}
-    )
-    
-    transformations = Compose([
-    {tf}
-    ])""".format(
-        pytorch2tensor=pytorch2tensor,
-        imports=res_imports,
-        tf=composes,
-    )
-
-    return autopep8.fix_code(result, options={
-        'max_line_length': 80,
-    })
